Include userId to firestore queries

+++ src/utils/firebase.ts
@@ -1,20 +1,42 @@
 import firestore, {
+  firebase,
   FirebaseFirestoreTypes,
 } from '@react-native-firebase/firestore';
 
+/**
+ * Las funciones que tienen el userId como parámetro opcional son las que pueden
+ * llegar a necesitar "autenticación", simplemente si se les pasa el parámetro
+ * se agrega el where(userId == ...)
+ */
 export interface IGenericCollection<Type> {
   collection?: FirebaseFirestoreTypes.CollectionReference<Type>;
   add: (data: Type) => Promise<string | null>;
-  get: (id: string) => Promise<Type | null>;
-  getAll: () => Promise<Array<Type> | null>;
-  update: (newData: Type & { id: string }) => Promise<boolean>;
-  remove: (id: string) => Promise<boolean>;
+  get: (id: string, userId?: string) => Promise<Type | null>;
+  getAll: (userId?: string) => Promise<Array<Type> | null>;
+  update: (newData: Type & { id: string }, userId?: string) => Promise<boolean>;
+  remove: (id: string, userId?: string) => Promise<boolean>;
 }
 
 export const getGenericCollection = <Type>(
   name: string,
 ): IGenericCollection<Type> => {
   const collection = firestore().collection<Type>(name);
+
+  const whereIdEqualsTo = (
+    id: string,
+    query?: FirebaseFirestoreTypes.Query<Type>,
+  ): FirebaseFirestoreTypes.Query<Type> =>
+    (query ?? collection).where(
+      firebase.firestore.FieldPath.documentId(),
+      '==',
+      id,
+    );
+
+  const whereUserIdEqualsTo = (
+    userId: string,
+    query?: FirebaseFirestoreTypes.Query<Type>,
+  ): FirebaseFirestoreTypes.Query<Type> =>
+    (query ?? collection).where('userId' as keyof Type, '==', userId);
 
   const add = async (data: Type): Promise<string | null> => {
     try {
@@ -27,10 +49,19 @@ export const getGenericCollection = <Type>(
     }
   };
 
-  const get = async (id: string): Promise<Type | null> => {
+  const get = async (id: string, userId?: string): Promise<Type | null> => {
     try {
-      const response = await collection.doc(id).get();
-      return response.data() ?? null;
+      let query = whereIdEqualsTo(id);
+      if (userId?.length) {
+        query = whereUserIdEqualsTo(userId, query);
+      }
+      const response = await query.get();
+      if (response.size === 1) {
+        const doc = response?.docs?.[0];
+        return doc?.id ? { ...doc.data(), id: doc.id } : null;
+      } else {
+        throw new Error(`Document does not exist or User is not the owner`);
+      }
     } catch (error) {
       console.error(`Error while fetching doc: ${id} from collection: ${name}`);
       console.error(`Error: ${error}`);
@@ -38,10 +69,15 @@ export const getGenericCollection = <Type>(
     }
   };
 
-  const getAll = async (): Promise<Array<Type> | null> => {
+  const getAll = async (userId?: string): Promise<Array<Type> | null> => {
     try {
-      const response = await collection.get();
-      return response.docs.map(doc => ({ id: doc.id, ...doc.data() }));
+      let response: FirebaseFirestoreTypes.QuerySnapshot<Type> | null = null;
+      if (userId?.length) {
+        response = await whereUserIdEqualsTo(userId).get();
+      } else {
+        response = await collection.get();
+      }
+      return response.docs.map(doc => ({ ...doc.data(), id: doc.id }));
     } catch (error) {
       console.error(`Error while fetching all docs from collection: ${name}`);
       console.error(`Error: ${error}`);
@@ -49,10 +85,18 @@ export const getGenericCollection = <Type>(
     }
   };
 
-  const update = async (newData: Type & { id: string }): Promise<boolean> => {
+  const update = async (
+    newData: Type & { id: string },
+    userId?: string,
+  ): Promise<boolean> => {
+    const document = await get(newData.id, userId);
     try {
-      await collection.doc(newData.id).update(newData);
-      return true;
+      if (document) {
+        await collection.doc(newData.id).update(newData);
+        return true;
+      } else {
+        throw new Error(`Document does not exist`);
+      }
     } catch (error) {
       console.error(
         `Error while updating doc: ${newData.id} in collection: ${name}`,
@@ -62,9 +106,12 @@ export const getGenericCollection = <Type>(
     }
   };
 
-  const remove = async (id: string): Promise<boolean> => {
+  const remove = async (id: string, userId?: string): Promise<boolean> => {
+    const document = await get(id, userId);
     try {
-      await collection.doc(id).delete();
+      if (document) {
+        await collection.doc(id).delete();
+      }
       return true;
     } catch (error) {
       console.error(`Error while deleting doc: ${id} from collection: ${name}`);
